backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
google-api-python-client==2.92.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
supabase==1.0.4
django-cors-headers==4.2.0
factory-boy==3.3.0
pytest==7.4.0
pytest-django==4.5.2